refactor: move BootConfig + parse_debug_watch into arch-abi
`BootConfig` (and the fw_cfg `opt/debug-watch` parser) is the boot boundary
contract — the platform entry constructs it, `startup` consumes it. It moves
from the kernel crate root into `arch-abi` so that, once the metal entry lives
in its own backend crate, both sides share one type. The kernel re-exports it,
so `crate::BootConfig` / `crate::parse_debug_watch` are unchanged. Both build.
